+++ services/instance-service/src/main/kotlin/org/katan/service/instance/http/routes/GetInstanceFsBucket.kt
@@ -8,9 +8,9 @@ import org.katan.http.response.HttpError
 import org.katan.http.response.respond
 import org.katan.http.response.respondError
 import org.katan.http.response.validateOrThrow
+import org.katan.model.instance.InstanceNotFoundException
 import org.katan.service.fs.FSService
 import org.katan.service.fs.http.dto.FSBucketResponse
-import org.katan.service.instance.InstanceNotFoundException
 import org.katan.service.instance.InstanceService
 import org.katan.service.instance.http.InstanceRoutes
 import org.koin.ktor.ext.inject
@@ -24,7 +24,7 @@ internal fun Route.getInstanceFsBucket() {
         validator.validateOrThrow(parameters)
 
         val instance = try {
-            instanceService.getInstance(parameters.instanceId.toLong())
+            instanceService.getInstance(parameters.instanceId)
         } catch (_: InstanceNotFoundException) {
             respondError(HttpError.UnknownInstance)
         }
