+++ services/instance-service/src/main/kotlin/org/katan/service/instance/http/routes/ReadFsFile.kt
@@ -12,8 +12,8 @@ import jakarta.validation.Validator
 import org.katan.http.response.HttpError
 import org.katan.http.response.respondError
 import org.katan.http.response.validateOrThrow
+import org.katan.model.instance.InstanceNotFoundException
 import org.katan.service.fs.FSService
-import org.katan.service.instance.InstanceNotFoundException
 import org.katan.service.instance.InstanceService
 import org.katan.service.instance.http.InstanceRoutes
 import org.koin.ktor.ext.inject
@@ -26,7 +26,7 @@ internal fun Route.readFsFile() {
     get<InstanceRoutes.FSReadFile> { parameters ->
         validator.validateOrThrow(parameters)
         val instance = try {
-            instanceService.getInstance(parameters.instanceId.toLong())
+            instanceService.getInstance(parameters.instanceId)
         } catch (_: InstanceNotFoundException) {
             respondError(HttpError.UnknownInstance)
         }
